--- a/src/app/global-error.tsx
+++ b/src/app/global-error.tsx
@@ -42,6 +42,7 @@ export default function GlobalError() {
           {/* Tautan untuk kembali ke halaman utama */}
           <Link
             href='/'
+            aria-label='Back to home'
             className='border-b-[1.5px] border-b-gray-800 font-satoshi text-xl font-medium transition-colors duration-150 hover:text-gray-400 md:text-xl lg:text-2xl'
           >
             &larr; Go back