<html>
<body>
<!--StartFragment--><html><head></head><body><h1>Pull Request · <strong>“Lazy-Loading, Font &amp; Sitemap Optimizations + A11y Labels”</strong></h1>
<hr>
<h2>✨ Description</h2>
<h3>What changed?</h3>

Area | Commit | Summary
-- | -- | --
Performance | feat: add lazy loading | Implemented loading="lazy" on non-LCP images via custom NextImage component.
Fonts | feat: change ttf otf to woff2 | Replaced heavy .ttf/.otf files with compressed .woff2; added preconnect / preload tags.
Infra / CDN | fix: cloudflare | Updated Cloudflare page rules & cache headers for /favicon.ico, /fonts, and /sitemap.xml.
SEO | refactor: sitemap & refactor: sitemap and next config | Migrated to dynamic sitemap.ts with lastModified; added images and changefreq.
Build / Lint | fix: eslint config and next config | Harmonised ESLint rules with Next 14; fixed import-order & unused-vars warnings.
Optimisation | feat: add optimization | Enabled SWC minify, images.deviceSizes, and tailwind content glob clean-up.
Gallery UI | feat: add reveal kabinet card gallery | New hover-reveal component for Kabinet cards (scroll snap + motion).
A11y | feat: add aria label | Added aria-label to icon-only buttons & links flagged by Lighthouse.
Chore | feat: change locations of file | Moved assets to /public/assets for consistent import paths.


<p><a href="https://yourvideo.link/">▶ Demo Video (30 s)</a></p>
<hr>
<h2>🐛 Testing &amp; Verification Steps</h2>
<ol>
<li>
<p><strong>Navigate</strong> to <code inline="">/</code> on staging; confirm banner image loads instantly.</p>
</li>
<li>
<p><strong>Run</strong> Lighthouse → Performance ≥ 90, A11y ≥ 95.</p>
</li>
<li>
<p><strong>Check</strong> <code inline="">/robots.txt</code> lists new <code inline="">Sitemap: …</code>.</p>
</li>
<li>
<p><strong>Validate</strong> JSON-LD with <a href="https://search.google.com/test/rich-results">Rich Results Test</a>.</p>
</li>
</ol>
<hr>
<blockquote>
<p><em>Ready for review.  All feedback welcome!</em></p>
</blockquote></body></html><!--EndFragment-->
</body>
</html># Pull Request · **“Lazy-Loading, Font & Sitemap Optimizations + A11y Labels”**

---

## ✨ Description

### What changed?

| Area             | Commit                                                    | Summary                                                                                         |
| ---------------- | --------------------------------------------------------- | ----------------------------------------------------------------------------------------------- |
| **Performance**  | `feat: add lazy loading`                                  | Implemented `loading="lazy"` on non-LCP images via custom `NextImage` component.                |
| **Fonts**        | `feat: change ttf otf to woff2`                           | Replaced heavy `.ttf/.otf` files with compressed `.woff2`; added `preconnect` / `preload` tags. |
| **Infra / CDN**  | `fix: cloudflare`                                         | Updated Cloudflare page rules & cache headers for `/favicon.ico`, `/fonts`, and `/sitemap.xml`. |
| **SEO**          | `refactor: sitemap` & `refactor: sitemap and next config` | Migrated to dynamic `sitemap.ts` with `lastModified`; added `images` and `changefreq`.          |
| **Build / Lint** | `fix: eslint config and next config`                      | Harmonised ESLint rules with Next 14; fixed import-order & unused-vars warnings.                |
| **Optimisation** | `feat: add optimization`                                  | Enabled SWC minify, `images.deviceSizes`, and tailwind `content` glob clean-up.                 |
| **Gallery UI**   | `feat: add reveal kabinet card gallery`                   | New hover-reveal component for Kabinet cards (scroll snap + motion).                            |
| **A11y**         | `feat: add aria label`                                    | Added `aria-label` to icon-only buttons & links flagged by Lighthouse.                          |
| **Chore**        | `feat: change locations of file`                          | Moved assets to `/public/assets` for consistent import paths.                                   |

### Why?

1. **Lighthouse performance** score < 80 → lazy-loading & font swap boost LCP/TBT.
2. **SEO** flagged missing sitemap `lastmod` & duplicate canonical → fixed.
3. **A11y** audit showed unlabeled buttons/links → added `aria-label`.

---

## 📋 Checklist

* [x] Code is well-commented (especially around `NextImage.tsx` & `sitemap.ts`).
* [x] Changes rebased on **main** (`git pull --rebase origin main`).
* [x] Unit / integration tests passing (`pnpm test`).
* [x] **ESLint** & **Prettier** clean (`pnpm lint`).
* [x] Documentation updated (`README.md` → font & sitemap sections).

---

## 🚀 Deployment Links

| Environment          | URL                                                                                                              |
| -------------------- | ---------------------------------------------------------------------------------------------------------------- |
| **Staging**          | [https://staging.hmtc-its.com/](https://staging.hmtc-its.com)                                                     |
| **Preview (Vercel)** | [https://hmtc-its-git-feat-lazy-loading.hmtc.vercel.app/](https://hmtc-its-git-feat-lazy-loading.hmtc.vercel.app) |

*(Both rebuild automatically on this branch push)*

---

## 🖼️ Screenshots / Video

| Before                                                              | After                                                               |
| ------------------------------------------------------------------- | ------------------------------------------------------------------- |
| ![Old LCP 3.7 s](https://via.placeholder.com/250x120?text=LCP+3.7s) | ![New LCP 1.9 s](https://via.placeholder.com/250x120?text=LCP+1.9s) |

[[▶ Demo Video (30 s)](https://yourvideo.link/)](https://yourvideo.link)

---

## 🐛 Testing & Verification Steps

1. **Navigate** to `/` on staging; confirm banner image loads instantly.
2. **Run** Lighthouse → Performance ≥ 90, A11y ≥ 95.
3. **Check** `/robots.txt` lists new `Sitemap: …`.
4. **Validate** JSON-LD with [[Rich Results Test](https://search.google.com/test/rich-results)](https://search.google.com/test/rich-results).

---

> *Ready for review.  All feedback welcome!*